Extract statistics package
