+++ src/statistics/none.go
@@ -1,0 +1,8 @@
+package statistics
+
+// NoStatistics is a statistics implementation that does nothing.
+type NoStatistics struct{}
+
+func (s *NoStatistics) RegisterRun() {}
+
+func (s *NoStatistics) PrintAnalysis() {}
